Extract Application into separate module
